[Chore] components 폴더 구조 세팅-common
